requirements.txt
asn1crypto==0.23.0
beautifulsoup4==4.6.0
bs4==0.0.1
certifi==2017.11.5
cffi==1.11.2
chardet==3.0.4
click==6.7
cryptography==2.1.3
Django==1.11.7
django-slack==5.10.2
djangorestframework==3.7.3
docx==0.2.4
enum34==1.1.6
Flask==0.12.2
idna==2.6
ipaddress==1.0.18
itsdangerous==0.24
Jinja2==2.10
lxml==3.8.0
MarkupSafe==1.0
olefile==0.44
Pillow==4.1.1
pycparser==2.18
pyee==3.0.3
pyOpenSSL==17.3.0
PyPDF2==1.26.0
pysolr==3.6.0
pytz==2017.3
requests==2.18.4
six==1.11.0
slackclient==1.0.9
slackeventsapi==1.0.0
urllib3==1.22
virtualenv==15.1.0
watson-developer-cloud==0.26.1
websocket-client==0.44.0
Werkzeug==0.12.2
